added apine schema, added bids2apine parser, added new examples, removed old scripts and examples
