--- scripts/flat_example.json
@@ -1,31 +1,0 @@
-{
-    "dataset_description": {
-        "BIDSVersion": "1.0.0",
-        "License": "This dataset is made available under the Public Domain Dedication and License \nv1.0, whose full text can be found at \nhttp://www.opendatacommons.org/licenses/pddl/1.0/. \nWe hope that all users will follow the ODC Attribution/Share-Alike \nCommunity Norms (http://www.opendatacommons.org/norms/odc-by-sa/); \nin particular, while not legally required, we hope that all users \nof the data will acknowledge the OpenfMRI project and NSF Grant \nOCI-1131441 (R. Poldrack, PI) in any publications.",
-        "Name": "Rhyme judgment"
-    },
-    "demographics": "/Users/greg/code/INCF/BIDS-examples/ds003/participants.tsv",
-    "README": "/Users/greg/code/INCF/BIDS-examples/ds003/README",
-    "CHANGES": "/Users/greg/code/INCF/BIDS-examples/ds003/CHANGES",
-    "queries": [
-      {
-        "key": "anat",
-        "objects": ["entity1", "entity2"]
-      }
-    ],
-    "entities": [
-      {
-        "entity": "/Users/greg/code/INCF/BIDS-examples/ds003/sub-01/anat/sub-01_inplaneT2.nii.gz",
-        "subj": "01",
-        "session": "test",
-        "modality": "anat",
-        "type": "inplaneT2",
-        "run": 1,
-        "metadata": {},
-        "events": "/path/to/events/file",
-        "task": "somethingFun",
-        "bval": "/path/to/bval/file",
-        "bvec": "/path/to/bvec/file"
-      }
-    ]
-}
